perf(app): optimize React rendering and query performance across components
- Add React.memo wrapping to TagTable, TagPageHeader, CompanyPageHeader, and MetadataProvider
- Implement useCallback/useMemo for event handlers and computed values
- Optimize React Query with request deduplication keys and notification props
- Add comprehensive logging with createLogger for performance monitoring
- Refactor ArtistList page with extracted ArtistPageHeader and ArtistTable components
- Improve query key structure and add user-scoped caching
- Remove unnecessary MetadataProvider wrappers in routing
- Add JSDoc documentation and author attribution across components
