+++ src/components/company/CompanyPageHeader.tsx
@@ -1,9 +1,24 @@
+/**
+ * Company Page Header Component
+ * @author @serabi
+ * @created 2025-01-09
+ */
+
+import React from 'react';
 import AddCompanyDialog from './AddCompanyDialog';
 
+/**
+ * Props interface for the CompanyPageHeader component
+ */
 interface CompanyPageHeaderProps {
   onCompanyAdded: () => void;
 }
 
+/**
+ * CompanyPageHeader Component
+ *
+ * Renders the header section for the Company page including title and "Add Company" dialog.
+ */
 const CompanyPageHeader = ({ onCompanyAdded }: CompanyPageHeaderProps) => {
   return (
     <div className="mb-8 flex flex-col items-start justify-between md:flex-row md:items-center">
@@ -17,4 +32,4 @@ const CompanyPageHeader = ({ onCompanyAdded }: CompanyPageHeaderProps) => {
   );
 };
 
-export default CompanyPageHeader;
+export default React.memo(CompanyPageHeader);
