+++ src/components/tags/TagPageHeader.tsx
@@ -1,9 +1,24 @@
+/**
+ * Tag Page Header Component
+ * @author @serabi
+ * @created 2025-01-09
+ */
+
+import React from 'react';
 import AddTagDialog from './AddTagDialog';
 
+/**
+ * Props interface for the TagPageHeader component
+ */
 interface TagPageHeaderProps {
   onTagAdded: () => void;
 }
 
+/**
+ * TagPageHeader Component
+ *
+ * Renders the header section for the Tag page including title and "Add Tag" dialog.
+ */
 const TagPageHeader = ({ onTagAdded }: TagPageHeaderProps) => {
   return (
     <div className="mb-8 flex flex-col items-start justify-between md:flex-row md:items-center">
@@ -17,4 +32,4 @@ const TagPageHeader = ({ onTagAdded }: TagPageHeaderProps) => {
   );
 };
 
-export default TagPageHeader;
+export default React.memo(TagPageHeader);
